+++ src/core/Constants.cs
@@ -7,6 +7,6 @@
 
         public const string General = "General";
         public const string Logs = "Logs";
-        public const string Tools = "Tools";
+        public const string Other = "Other";
     }
 }