+++ src/options/Constants/Guids.cs
@@ -10,7 +10,7 @@ namespace ExtensibilityLogs.Options
         internal const string LogsDialogPageString = "5b42aa24-e633-48e2-9d2e-fbdb0494f36e";
         public static Guid LogsDialogPage = new Guid(LogsDialogPageString);
 
-        internal const string ToolsDialogPageString = "edd0f91a-df78-46e0-8479-4897104a127d";
-        public static Guid ToolsDialogPage = new Guid(ToolsDialogPageString);
+        internal const string OtherDialogPageString = "edd0f91a-df78-46e0-8479-4897104a127d";
+        public static Guid OtherDialogPage = new Guid(OtherDialogPageString);
     }
 }