+++ src/vsix/Commands/Other/EnvironmentVariablesCommand.cs
@@ -6,9 +6,9 @@ using System;
 using System.Collections;
 using static System.Environment;
 
-namespace ExtensibilityLogs.Commands.Tools
+namespace ExtensibilityLogs.Commands.Other
 {
-    internal sealed class EnvironmentVariablesCommand : ToolsCommand
+    internal sealed class EnvironmentVariablesCommand : OtherCommand
     {
         private EnvironmentVariablesCommand(PackageBase package) : base(package, PackageIds.EnvironmentVariablesCommand)
         { }
@@ -17,7 +17,7 @@ namespace ExtensibilityLogs.Commands.Tools
             => Instantiate(new EnvironmentVariablesCommand(package));
 
         protected override bool CanExecute
-            => base.CanExecute && PackageClass.ToolsOptions.EnvironmentVariablesCommandEnabled;
+            => base.CanExecute && PackageClass.OtherOptions.EnvironmentVariablesCommandEnabled;
 
         protected override void OnExecute(OleMenuCommand command)
 #pragma warning disable VSTHRD010 // Invoke single-threaded types on Main thread